get rid of remnants of iCountFragNL that was for multiple fragment neutral loss support; add casts to suppress Visual Studio warnings.
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@1476 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ branches/release_2020010/CometSearch/CometDataInternal.h
@@ -364,7 +364,7 @@ struct DBIndex
 
          // masses are the same at this point
          // next compare modification states
-         int iLen = strlen(szPeptide)+2;
+         int iLen = (int)strlen(szPeptide)+2;
          for (int i=0; i<iLen; i++)
          {
             if (pcVarModSites[i] != rhs.pcVarModSites[i])
